refactored inspection and AST parsing code, bump to version 0.5.0
