+++ py2puml/inspection/inspectmodule.py
@@ -1,0 +1,53 @@
+
+from typing import Dict, Iterable, List, Type
+from types import ModuleType
+
+from enum import Enum
+from inspect import getmembers, isclass
+
+from py2puml.domain.umlitem import UmlItem
+from py2puml.domain.umlrelation import UmlRelation
+from py2puml.inspection.inspectclass import inspect_class_type
+from py2puml.inspection.inspectenum import inspect_enum_type
+from py2puml.inspection.inspectnamedtuple import inspect_namedtuple_type
+
+def filter_domain_definitions(module: ModuleType, root_module_name: str) -> Iterable[Type]:
+    for definition_key in dir(module):
+        definition_type = getattr(module, definition_key)
+        if isclass(definition_type):
+            definition_members = getmembers(definition_type)
+            definition_module_member = next((
+                member for member in definition_members
+                # ensures that the type belongs to the module being parsed
+                if member[0] == '__module__' and member[1].startswith(root_module_name)
+            ), None)
+            if definition_module_member is not None:
+                yield definition_type
+
+def inspect_type(
+    definition_type: Type,
+    root_module_name: str,
+    domain_items_by_fqdn: Dict[str, UmlItem],
+    domain_relations: List[UmlRelation]
+):
+    definition_type_fqdn = f'{definition_type.__module__}.{definition_type.__name__}'
+    if definition_type_fqdn not in domain_items_by_fqdn:
+        if issubclass(definition_type, Enum):
+            inspect_enum_type(definition_type, definition_type_fqdn, domain_items_by_fqdn)
+        elif getattr(definition_type, '_fields', None) is not None:
+            inspect_namedtuple_type(definition_type, definition_type_fqdn, domain_items_by_fqdn)
+        else:
+            inspect_class_type(
+                definition_type, definition_type_fqdn,
+                root_module_name, domain_items_by_fqdn, domain_relations
+            )
+
+def inspect_module(
+    domain_item_module: ModuleType,
+    root_module_name: str,
+    domain_items_by_fqdn: Dict[str, UmlItem],
+    domain_relations: List[UmlRelation]
+):
+    # processes only the definitions declared or imported within the given root module
+    for definition_type in filter_domain_definitions(domain_item_module, root_module_name):
+        inspect_type(definition_type, root_module_name, domain_items_by_fqdn, domain_relations)
