+++ py2puml/parsing/compoundtypesplitter.py
