+++ py2puml/py2puml.py
@@ -5,7 +5,7 @@ from importlib import import_module
 
 from py2puml.domain.umlitem import UmlItem
 from py2puml.domain.umlrelation import UmlRelation
-from py2puml.parser import parse_item_module
+from py2puml.inspection.inspectmodule import inspect_module
 from py2puml.exportpuml import to_puml_content
 
 
@@ -15,7 +15,7 @@ def py2puml(domain_path: str, domain_module: str) -> Iterable[str]:
     for _, name, is_pkg in walk_packages([domain_path], f'{domain_module}.'):
         if not is_pkg:
             domain_item_module: ModuleType = import_module(name)
-            parse_item_module(
+            inspect_module(
                 domain_item_module,
                 domain_module,
                 domain_items_by_fqdn,
