+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "py2puml"
-version = "0.4.0"
+version = "0.5.0"
 description = "Generate Plantuml diagrams to document your python code"
 keywords = ["class diagram", "PlantUML", "documentation"]
 readme = "README.md"
