+++ tests/modules/withenum.py
@@ -1,3 +1,4 @@
+
 from enum import Enum
 
 class TimeUnit(Enum):
