+++ tests/py2puml/test__init__.py
@@ -2,7 +2,7 @@ from tests import __version__, __description__
 
 # Ensures the library version is modified in the pyproject.toml file when upgrading it (pull request)
 def test_version():
-    assert __version__ == '0.4.0'
+    assert __version__ == '0.5.0'
 
 # Description also output in the CLI
 def test_description():
